fuck a duck triangle transformation finally working

+++ as4/classes.h
@@ -90,7 +90,7 @@ public:
 	// methods
 	void translate(vec3 t);
 	void translate(float x, float y, float z) ;
-	void rotate(int angle, vec3 u);
+	void rotate(float angle, vec3 u);
 	void scale(float xScale, float yScale, float zScale);
 	void scale(vec3 s);
 	virtual vec4 normal(vec4)=0;
